fix: bump nu-protocol to v0.92
